Fix gcc warnings.
This is just to make -Werror succeed as the samples are not that
important.

+++ libnet/sample/udp1.c
@@ -198,7 +198,8 @@ main(int argc, char *argv[])
     fprintf(stderr, "Packets sent:  %lld\n"
                     "Packet errors: %lld\n"
                     "Bytes written: %lld\n",
-                    ls.packets_sent, ls.packet_errors, ls.bytes_written);
+                    (long long)ls.packets_sent, (long long)ls.packet_errors,
+		    (long long)ls.bytes_written);
     libnet_destroy(l);
     return (EXIT_SUCCESS);
 bad:
